clean up

+++ src/main/java/org/medicmobile/webapp/mobile/NotificationForegroundHandler.java
@@ -8,8 +8,8 @@ import android.webkit.WebView;
 public class NotificationForegroundHandler {
 	public static final String DEBUG_TAG = "NOTIFICATION_HANDLER";
 	private final Handler handler = new Handler(Looper.getMainLooper());
-	private static final long INTERVAL_MILLIS = 5 * 60 * 1000; //5mins interval
-	private static final long INITIAL_EXECUTION_DELAY = NotificationWorker.EXECUTION_TIMEOUT_SECS * 1000;
+	private static final int INTERVAL_MILLIS = 5 * 60 * 1000; //5mins interval
+	private static final int INITIAL_EXECUTION_DELAY = NotificationWorker.EXECUTION_TIMEOUT_SECS * 1000;
 
 	private final Runnable runnable;
 	private boolean isRunning = false;
